Allow for a catch-all bin for atom types during graph construction
In the construction of graphs from structures and molecules, a list
of allowed elements is provided. However, if an atom is provided in
a structure not found in the provided lists, a Value or KeyError is
raised.

This commit allows the user to set an argument `allow_other_atoms`
during instantiation of `Structure2Graph` and `Molecule2Graph`, which
if True, treats any atom not found in the provided list as its own
category.

For example, if `element_types=["Ti", "O"]` but a structure contains
Ti, O and Mg, then Ti will be treated by index 0, O will be treated
by index 1 and Mg will be treated by index `len(element_types)`,
which in this case is 2.