Update Training a MEGNet Formation Energy Model with PyTorch Lightning.ipynb (#185)
Bug fix: Resolved parameter input issue in MEGNetDataset from v0.8.3 and higher versions
Description:
This pull request addresses the problem of parsing input parameters when constructing MEGNetDataset in this scenario.  The issue arose due to the refactoring of MEGNetDataset in v0.8.3 and higher versions.  The modifications made here ensure compatibility with the syntax of the new version.

Signed-off-by: Ming <55485031+1152041831@users.noreply.github.com>