Ensure `free_energy` property is a `float` and not `np.array(float)`
The ASE `Calculator.results` attribute of `free_energy` should be formatted as a `float` but is currently formatted as a `np.array(float)`. For instance, `array(-4.0938973, dtype=float32)`. Since this is not the expected ASE type and because it is not iterable, this can cause a variety of unexpected problems. https://github.com/materialsvirtuallab/monty/issues/668 is just one example.

I have ensured `free_energy` is returned as a `float` in this PR.

Signed-off-by: Andrew S. Rosen <asrosen93@gmail.com>